feat: altera lógica dos Objetivos e adicionar data de início e fim

+++ src/main/java/br/com/codemain/nutrixpertai/dto/food/FoodResponseDTO.java
@@ -4,7 +4,7 @@ public record FoodResponseDTO(
         Long id,
         Long mealId,
         String foodName,
-        Double calories,
+        Integer calories,
         Double protein,
         Double carbohydrates,
         Double fat) {
