+++ src/main/java/br/com/codemain/nutrixpertai/dto/food/UpdateFoodDTO.java
@@ -10,7 +10,7 @@ public record UpdateFoodDTO(
 
         @NotNull(message = "Calorias são obrigatórias")
         @Positive(message = "Calorias devem ser positivas")
-        Double calories,
+        Integer calories,
 
         @NotNull(message = "Proteínas são obrigatórias")
         @Positive(message = "Proteínas devem ser positivas")
